suite/common/places/tests/chrome/test_0_bug510634.xul

<?xml version="1.0"?>

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->

<?xml-stylesheet href="chrome://global/skin" type="text/css"?>
<?xml-stylesheet href="chrome://mochikit/content/tests/SimpleTest/test.css"
                 type="text/css"?>

<?xml-stylesheet href="chrome://communicator/content/places/places.css"?>
<?xml-stylesheet href="chrome://communicator/skin/places/places.css"?>
<?xul-overlay href="chrome://communicator/content/bookmarks/placesOverlay.xul"?>

<window xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        title="510634: Wrong icons on bookmarks sidebar"
        onload="runTest();">

  <script type="application/javascript"
          src="chrome://mochikit/content/MochiKit/packed.js" />
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js" />

  <body xmlns="http://www.w3.org/1999/xhtml" />

  <tree id="tree"
        type="places"
        flex="1">
    <treecols>
      <treecol label="Title" id="title" anonid="title" primary="true" ordinal="1" flex="1"/>
    </treecols>
    <treechildren flex="1"/>
  </tree>

  <script type="application/javascript">
  <![CDATA[

    /**
     * Bug 510634 -  Wrong icons on bookmarks sidebar
     * https://bugzilla.mozilla.org/show_bug.cgi?id=510634
     *
     * Ensures that properties for special queries are set on their tree nodes,
     * even if PlacesUIUtils.leftPaneFolderId was not initialized.
     */

    SimpleTest.waitForExplicitFinish();

    function createSupportsArray() {
      return Components.classes["@mozilla.org/supports-array;1"]
                       .createInstance(Components.interfaces.nsISupportsArray);
    }

    // converts nsISupportsArray of atoms to a simple JS-strings array
    function convertPropertiesToJSArray(aSupportsArray) {
      var results = [];
      var count = aSupportsArray.Count();
      for (var i = 0; i < count; i++)
        results.push(aSupportsArray.QueryElementAt(i, Components.interfaces.nsIAtom).toString());

      return results;
    }

    function runTest() {
      // We need to cache and restore this getter in order to simulate
      // Bug 510634
      let cachedLeftPaneFolderIdGetter =
        PlacesUIUtils.__lookupGetter__("leftPaneFolderId");

      let leftPaneFolderId = PlacesUIUtils.leftPaneFolderId;

      // restore the getter
      PlacesUIUtils.__defineGetter__("leftPaneFolderId", cachedLeftPaneFolderIdGetter);

      // Setup the places tree contents.
      let tree = document.getElementById("tree");
      tree.place = "place:queryType=1&folder=" + leftPaneFolderId;

      // Open All Bookmarks
      PlacesUtils.asContainer(tree.view.nodeForTreeIndex(1)).containerOpen = true;

      // The query-property is set on the title column for each row.
      let titleColumn = tree.treeBoxObject.columns.getColumnAt(0);

      ["Tags", "AllBookmarks", "BookmarksToolbar",
       "BookmarksMenu", "UnfiledBookmarks"].forEach(
         function(aQueryName, aRow) {
           let rowProperties = createSupportsArray();
           tree.view.getCellProperties(aRow, titleColumn, rowProperties);
           rowProperties = convertPropertiesToJSArray(rowProperties);
           ok(rowProperties.indexOf("OrganizerQuery_" + aQueryName) != -1,
             "OrganizerQuery_" + aQueryName + " is set");
         }
       );

      // Close the root node
      tree.result.root.containerOpen = false;

      // Restore the getter for the next test.
      PlacesUIUtils.__defineGetter__("leftPaneFolderId", cachedLeftPaneFolderIdGetter);

      SimpleTest.finish();
    }

  ]]>
  </script>
</window>
